chore: convert quickstarts to the Timefold Models Service SDK
Convert the 15 use-case quickstarts for platform/SDK compatibility:
build on the timefold-solver-service-parent, remove enterprise edition
dependencies, refactor scoring to HardMediumSoftScore, expand test
coverage, and add PMD/CPD checks. Point the CI workflows at the real
use-cases/* module paths and drop the school-timetabling Gradle build,
which the SDK conversion no longer supports. getting-started/* is
untouched.

+++ use-cases/task-assigning/src/main/java/org/acme/taskassigning/domain/Task.java
@@ -46,7 +46,7 @@ public class Task {
     }
 
     public Task(String id, TaskType taskType, int indexInTaskType, Customer customer, Employee employee, int minStartTime,
-                Priority priority) {
+            Priority priority) {
         this.id = id;
         this.taskType = taskType;
         this.indexInTaskType = indexInTaskType;
@@ -133,12 +133,12 @@ public class Task {
     // ************************************************************************
 
     @SuppressWarnings("unused")
-    @ShadowSources({"employee", "previousTask.startTime"})
+    @ShadowSources({ "employee", "previousTask.startTime" })
     public Long startTimeSupplier() {
         if (employee == null) {
             return null;
         } else if (previousTask == null) {
-            return (long) minStartTime;
+            return minStartTime;
         } else {
             var previousEndTime = previousTask.getEndTime();
             return Math.max(previousEndTime, minStartTime);
